v2/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "privategpt"
version = "0.2.0"
description = "PrivateGPT v2 monorepo"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "PrivateGPT Team"}]

# Core runtime dependencies
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sqlalchemy>=2.0",
    "bcrypt>=4.0",
    "python-jose[cryptography]>=3.3",
    "pydantic>=2.6",
    "httpx>=0.27",
    "structlog>=24.1",
    "email-validator>=2.0",
    "numpy>=1.26",
    "pytest-asyncio>=0.23",
    "aiosqlite>=0.19",
    "weaviate-client>=3.25",
    "sentence-transformers>=2.6",
    "torch>=2.1; platform_machine!='arm64' or sys_platform!='darwin'",
    "torch==2.1.2; platform_machine=='arm64' and sys_platform=='darwin'",
    "python-multipart>=0.0.7",
    "celery>=5.3",
    "pydantic-settings>=2.1"
]

[tool.setuptools.packages.find]
where = ["src"]
